+++ backend/core/src/seeder/seed.ts
@@ -28,6 +28,7 @@ import {
   getPbvPreference,
   getServedInMilitaryProgram,
   getTayProgram,
+  getFlatRentAndRentBasedOnIncomeProgram,
 } from "./seeds/listings/shared"
 import { UserCreateDto } from "../auth/dto/user-create.dto"
 import { AmiDefaultSanJose } from "./seeds/ami-charts/default-ami-chart-san-jose"
@@ -153,6 +154,7 @@ export async function createPrograms(app: INestApplicationContext, jurisdictions
     getTayProgram(),
     getDisabilityOrMentalIllnessProgram(),
     getHousingSituationProgram(),
+    getFlatRentAndRentBasedOnIncomeProgram(),
   ])
 
   for (const jurisdiction of jurisdictions) {
@@ -259,6 +261,31 @@ async function seed() {
     new AuthContext(null)
   )
   await userService.confirm({ token: user2.confirmationToken })
+
+  // create user with expired password
+  const userExpiredPassword = await userService.createPublicUser(
+    plainToClass(UserCreateDto, {
+      email: "user+expired@example.com",
+      emailConfirmation: "user+expired@example.com",
+      firstName: "Second",
+      middleName: "Mid",
+      lastName: "Last",
+      dob: new Date(),
+      password: "abcdef",
+      passwordConfirmation: "abcdef",
+      jurisdictions: [jurisdictions[0]],
+      roles: { isAdmin: false, isPartner: true },
+    }),
+    new AuthContext(null)
+  )
+
+  await userService.confirm({ token: userExpiredPassword.confirmationToken })
+
+  userExpiredPassword.passwordValidForDays = 180
+  userExpiredPassword.passwordUpdatedAt = new Date("2020-01-01")
+  userExpiredPassword.confirmedAt = new Date()
+
+  await userRepo.save(userExpiredPassword)
 
   const admin = await userService.createPublicUser(
     plainToClass(UserCreateDto, {